Build pipeline tuning

+++ Source/Common/DiskFileSystem.cpp
@@ -148,7 +148,7 @@ auto DiskFile::Read(void* buf, size_t len) -> bool
     }
 
     DWORD br = 0;
-    return ::ReadFile(_pImpl->FileHandle, buf, len, &br, nullptr) != 0 && br == len;
+    return ::ReadFile(_pImpl->FileHandle, buf, len, &br, nullptr) != 0 && static_cast<size_t>(br) == len;
 }
 
 auto DiskFile::Write(const void* buf, size_t len) -> bool
@@ -163,7 +163,7 @@ auto DiskFile::Write(const void* buf, size_t len) -> bool
     }
 
     DWORD bw = 0;
-    return ::WriteFile(_pImpl->FileHandle, buf, static_cast<DWORD>(len), &bw, nullptr) != 0 && bw == len;
+    return ::WriteFile(_pImpl->FileHandle, buf, static_cast<DWORD>(len), &bw, nullptr) != 0 && static_cast<size_t>(bw) == len;
 }
 
 auto DiskFile::SetPos(int offset, DiskFileSeek origin) -> bool
